kPASLastFmTimeoutInSec set to 10

+++ Passions/PASResources.h
@@ -14,7 +14,7 @@ extern NSString * const kPASParseClientKey;
 extern NSString * const kPASParseMasterKey;
 
 extern NSString * const kPASLastFmApiKey;
-extern CGFloat const kPASLastFmTimeoutInSec;
+extern float const kPASLastFmTimeoutInSec;
 
 extern NSString * const kPASSpotifyClientId;
